vector elem size is now up to 64bit value, and elem count is up to 64bit value

+++ structures/node_list.c
@@ -10,7 +10,7 @@ node_list* create_nodelist(size_t elem_size){
     l->root = NULL;
 }
 
-size_t get_size(node_list* l){
+long long get_size(node_list* l){
     return l->elem_count;
 }
